+++ samples/job_meter/externs.h
@@ -1,3 +1,10 @@
+// Auto-generated, do not edit
+#ifndef EXTERNS_H
+#define EXTERNS_H
+
+#include "out/platform.h"
 #include "out/charInclude.h"
 #include "out/fixData.h"
 #include "out/gfx_dat.h"
+
+#endif