+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## 3.4.4
   - Externalized CDDA functions in preparation for integrating Neo-Geo Cartridge builds.  
     To use CDDA functions, you must now include `<cdda.h>`.
+    Checkout CDDA in samples folder
   - Added auto-generated `out/platform.h` to define the platform type (CD or Cartridge).  
     *Note: only the CD platform is supported for now.*
     Checkout CDDA in samples folder